pyproject.toml
[tool.poetry]
name = "coursedrf"
version = "0.1.0"
description = ""
authors = ["Aptemond <sofartem5@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.7"
python-decouple = "^3.8"
psycopg2 = "^2.9.9"
pillow = "^10.1.0"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.3.0"
django-filter = "^23.3"
coverage = "^7.3.2"
drf-yasg = "^1.21.7"
django-cors-headers = "^4.3.0"
requests = "^2.31.0"
celery = "^5.3.4"
eventlet = "^0.33.3"
django-celery-beat = "^2.5.0"
redis = "^5.0.1"
pytelegrambotapi = "^4.14.0"
python-telegram-bot-calendar = "^1.0.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
